+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Update Lightsheet directory schema
 - Update Histology to include description on OME-TIFFs
 - Update Histology with links
+- Disallow deprecated Organ TSV spec from validating
 
 ## v0.0.17